@@ -14,19 +14,20 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "axios": "^0.25.0",
     "crypto": "^1.0.1",
     "discord-oauth2": "^2.9.0",
-    "firebase-admin": "^9.8.0",
+    "firebase-admin": "^10.0.2",
     "firebase-functions": "^3.14.1"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^3.9.1",
-    "@typescript-eslint/parser": "^3.8.0",
-    "eslint": "^7.6.0",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
+    "eslint": "^8.7.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-import": "^2.22.0",
-    "firebase-functions-test": "^0.2.0",
-    "typescript": "^3.8.0"
+    "firebase-functions-test": "^0.3.3",
+    "typescript": "^4.5.5"
   },
   "private": true
 }
