@@ -1,8 +1,8 @@
 import {https, config} from "firebase-functions";
 import {Response} from "express";
 import {modifyResponse, isAuthenticated} from "../../auth/authentication";
-// import fetch = require("node-fetch");
-
+import axios from "axios";
+import {Agent} from "https";
 export const getServers = https
     .onRequest(async (request: https.Request, response: Response) => {
       if (request.method === "OPTIONS") {
@@ -15,17 +15,19 @@ export const getServers = https
       }
       response = modifyResponse(request, response, 200);
 
-      const base64BasicAuth = new Buffer(
+      const base64BasicAuth = Buffer.from(
           `${config().nodeone.user+":"+config().nodeone.password}`
       ).toString("base64");
-      const nodeOneUri = `${config().nodeone.ip}:${config().nodeone.port}`;
-      const fetched = await fetch(`${nodeOneUri}/servers`, {
+      const nodeOneUri = `https://${config().nodeone.ip}:${config().nodeone.port}`;
+      const fetched = await axios.request({
+        url: `${nodeOneUri}/servers`,
         method: "GET",
         headers: {
           authorization: `Basic ${base64BasicAuth}`,
         },
+        httpsAgent: new Agent({rejectUnauthorized: false}),
       });
-      response.json(fetched.json());
+      response.json(fetched.data);
       return;
     });
 
@@ -41,16 +43,18 @@ export const getServer = https
       }
       response = modifyResponse(request, response, 200);
 
-      const base64BasicAuth = new Buffer(
+      const base64BasicAuth = Buffer.from(
           `${config().nodeone.user+":"+config().nodeone.password}`
       ).toString("base64");
-      const nodeOneUri = `${config().nodeone.ip}:${config().nodeone.port}`;
-      const fetched = await fetch(`${nodeOneUri}/server/${request.params.id}`, {
-        method: "GET",
+      const nodeOneUri = `https://${config().nodeone.ip}:${config().nodeone.port}`;
+      const fetched = await axios.request({
+        url: `${nodeOneUri}/server/${request.params.id}`,
+        method: "POST",
         headers: {
           authorization: `Basic ${base64BasicAuth}`,
         },
+        httpsAgent: new Agent({rejectUnauthorized: false}),
       });
-      response.json(fetched.json());
+      response.json(fetched.data);
       return;
     });
